requirements.txt
scikit-learn==1.3.2
matplotlib==3.8.2
cmake==3.28.1
buildtools==1.0.6
ecos==2.0.12
tensorflow-2.15.0